requirements.txt
fastapi
uvicorn
pydantic
pandas
pyarrow
numpy
faiss-cpu
sentence-transformers
torch
requests
streamlit